refactor: move bwrap config helpers into dedicated module
